[Major refactoring] Renamed most methods and interfaces for coherency and subtility.
This program was previously a technical recruitment test that would
only parse given html pages. I modified it so it works with webpages,
although in a limited way.
For obvious reasons, I have changed names of methods and interfaces
so it does not get found out for the wrong reasons.